requirements.txt
openai 
python-dotenv
ipykernel
pydantic-ai
langchain
langchain-openai